requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alabaster==0.7.16
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cuda-python==12.6.0
docopt==0.6.2
docutils==0.20.1
frozenlist==1.4.1
gevent==24.2.1
geventhttpclient==2.0.2
greenlet==3.0.3
grpcio==1.66.1
idna==3.8
imagesize==1.4.1
indic-nlp-library==0.92
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mock==5.1.0
Morfessor==2.0.6
mosestokenizer==1.2.1
multidict==6.0.5
nltk==3.9.1
numpy==1.26.4
nvidia-pyindex==1.0.9
openfile==0.0.7
packaging==24.1
pandas==2.2.2
protobuf==4.25.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-rapidjson==1.20
pytz==2024.1
regex==2024.7.24
requests==2.32.3
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinx-argparse==0.5.2
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
subword-nmt==0.3.8
tomli==2.0.1
toolwrapper==2.1.0
tqdm==4.66.5
tritonclient==2.49.0
tzdata==2024.1
uctools==1.3.0
urllib3==2.2.2
yarl==1.9.7
zope.event==5.0
zope.interface==7.0.3
